Refactor PhatRPC + PhatClient -- cache not needed for next release
As the cache system wasn't working yet and won't be needed for some
time, I pulled the code out. We can pull it back in once we're confident
of the base system.

PhatClient tests were moved back into the same directory.
By default, tests live with the code they're testing in Go.